cpp lint

+++ src/server/redis_connection.h
@@ -77,7 +77,7 @@ class Connection : public EvbufCallbackBase<Connection> {
   // https://github.com/redis/redis/blob/7.2/src/networking.c#L1099
   std::string VerbatimString(std::string ext, const std::string &data) const {
     CHECK(ext.size() <= 3);
-    int padded_len = 3 - ext.size();
+    int padded_len = 3 - static_cast<int>(ext.size());
     ext = ext + std::string(padded_len, ' ');
     return protocol_version_ == RESP::v3 ? "=" + std::to_string(3 + data.size() + 1) + CRLF + ext + ":" + data + CRLF
                                          : BulkString(data);
